Implemented doubles

+++ Code/PID Controller/lib/Controllers/Controler.h
@@ -4,10 +4,10 @@
 // Include libraries
 #include <avr/io.h>
 #include <util/delay.h>
-float e[3];
-float u[3];
+double e[3];
+double u[3];
 
 // Declare functions
-float control(float ref, float mesurement);
+double control(double ref, double mesurement);
 void step(uint8_t base, uint8_t top, uint16_t delay);
 #endif /* CONTROLER_H_ */
